pyproject.toml
[project]
name = "telegram-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "python-telegram-bot>=22.3",
    "requests>=2.32.5",
    "rich>=14.1.0",
]
